# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Astudio Backend Assessment (Laravel)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timesheets": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get list of timesheets",
                "description": "Returns list of timesheets",
                "operationId": "getTimesheetsList",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter timesheets by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filter timesheets by project ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Create a new timesheet",
                "description": "Store a newly created timesheet in storage",
                "operationId": "storeTimesheet",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task_name",
                                    "date",
                                    "hours",
                                    "user_id",
                                    "project_id"
                                ],
                                "properties": {
                                    "task_name": {
                                        "type": "string",
                                        "example": "Development Task"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-10"
                                    },
                                    "hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "user_id": {
                                        "description": "The ID of the user logging the timesheet",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "project_id": {
                                        "description": "The ID of the project related to the timesheet",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timesheet created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "task_name": {
                                            "type": "string",
                                            "example": "Development Task"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "hours": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "project_id": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/timesheets/{id}": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get a timesheet by ID",
                "description": "Display the specified timesheet",
                "operationId": "showTimesheet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the timesheet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "task_name": {
                                            "type": "string",
                                            "example": "Development Task"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "hours": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "project_id": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/timesheets/update": {
            "put": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Update an existing timesheet",
                "description": "Update the specified timesheet in storage",
                "operationId": "updateTimesheet",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "task_name": {
                                        "type": "string",
                                        "example": "Updated Task"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-10"
                                    },
                                    "hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Timesheet updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "task_name": {
                                            "type": "string",
                                            "example": "Updated Task"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "hours": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "project_id": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/timesheets/delete": {
            "delete": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Delete a timesheet",
                "description": "Remove the specified timesheet from storage",
                "operationId": "deleteTimesheet",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Sanctum token prefixed with \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}